requirements.txt
flask
googlesearch-python
newspaper3k